Rename the centerline module
